Complete API contract validation and address data type issues
Update SPRING_BOOT_MIGRATION_SPECIFICATION.md to reflect completion of API contract validation (Task 12.1), resolving compilation errors. Task 12.2 (Data Integrity Verification) is now blocked due to schema validation issues, specifically INTEGER/BIGINT type mismatches in the Devotee entity and related services. Includes fixes for `appointedBy` field type, addition of `HashMap` import, and updates to `Devotee.java` and `DtoMapper.java`. Also updates `run-spring-boot.sh` with a DATABASE_URL.

+++ spring-boot-backend/src/main/java/com/namhatta/model/entity/Devotee.java
@@ -128,7 +128,7 @@ public class Devotee extends BaseEntity {
     private String appointedDate;
     
     @Column(name = "appointed_by")
-    private Long appointedBy;
+    private Integer appointedBy;
     
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "appointed_by", insertable = false, updatable = false)
